Untested but theoretically finished version of the vSphere Api
